fix -Wshorten-64-to-32 warning(s)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38219 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/pet/pet.c
@@ -692,7 +692,8 @@ int machine_specific_init(void)
 
     /* Initialize sound.  Notice that this does not really open the audio
        device yet.  */
-    sound_init(machine_timing.cycles_per_sec, machine_timing.cycles_per_rfsh);
+    sound_init((unsigned int)machine_timing.cycles_per_sec,
+               (unsigned int)machine_timing.cycles_per_rfsh);
 
     /* Initialize keyboard buffer.  FIXME: Is this correct?  */
     /* moved to mem_load() because it's model specific... AF 30jun1998
